Cleanup internals
