--- src/vanilla/TupleAndReferences.tsx
@@ -1,3 +1,0 @@
-import { AnyScopeTuple } from "./types";
-
-
