+++ src/vanilla/internal/weakCache.test.ts
@@ -1,16 +1,17 @@
 import { createDeepCache } from "./weakCache";
 
 describe("Weak cache", () => {
-  it("order matters", () => {
-    const memoize = createDeepCache();
 
-    const first = {};
-    const second = {};
-    let i = 0;
-    const fn = () => ({
-      value: i++,
-    });
+  const memoize = createDeepCache();
 
+  const first = {};
+  const second = {};
+  let i = 0;
+  const fn = () => ({
+    value: i++,
+  });
+
+  it("Caches based on orders of dependencies", () => {
     // Same order, so same value
     const firstValue = memoize.deepCache(fn, [first, second]);
     const secondValue = memoize.deepCache(fn, [first, second]);
